Create a globalOpts struct to hold values common across commands
