+++ internal/pkg/cli/cli.go
@@ -4,10 +4,57 @@
 // Package cli contains the archer subcommands.
 package cli
 
+import (
+	"fmt"
+
+	"github.com/aws/amazon-ecs-cli-v2/internal/pkg/workspace"
+	"github.com/spf13/viper"
+)
+
+func init() {
+	// Store the local project's name in viper.
+	bindProjectName()
+}
+
+// globalOpts holds fields that are used across multiple commands.
+type globalOpts struct {
+	projectName string
+}
+
 // actionCommand is the interface that every command that creates a resource implements.
 type actionCommand interface {
 	Ask() error
 	Validate() error
 	Execute() error
 	RecommendedActions() []string
+}
+
+// bindProjectName loads the project's name to viper.
+// If there is an error, we swallow the error and leave the default value as empty string.
+func bindProjectName() {
+	name, err := loadProjectName()
+	if err != nil {
+		return
+	}
+	viper.SetDefault(projectFlag, name)
+}
+
+// loadProjectName retrieves the project's name from the workspace if it exists and returns it.
+// If there is an error, it returns an empty string and the error.
+func loadProjectName() (string, error) {
+	// Load the workspace and set the project flag.
+	ws, err := workspace.New()
+	if err != nil {
+		// If there's an error fetching the workspace, fall back to requiring
+		// the project flag be set.
+		return "", fmt.Errorf("fetching workspace: %w", err)
+	}
+
+	summary, err := ws.Summary()
+	if err != nil {
+		// If there's an error reading from the workspace, fall back to requiring
+		// the project flag be set.
+		return "", fmt.Errorf("reading from workspace: %w", err)
+	}
+	return summary.ProjectName, nil
 }
