+++ internal/pkg/cli/env.go
@@ -4,8 +4,6 @@
 package cli
 
 import (
-	"fmt"
-
 	"github.com/aws/amazon-ecs-cli-v2/internal/pkg/archer"
 	"github.com/aws/amazon-ecs-cli-v2/internal/pkg/deploy"
 	"github.com/spf13/cobra"
@@ -13,7 +11,6 @@ import (
 
 	"github.com/aws/amazon-ecs-cli-v2/cmd/archer/template"
 	"github.com/aws/amazon-ecs-cli-v2/internal/pkg/cli/group"
-	"github.com/aws/amazon-ecs-cli-v2/internal/pkg/workspace"
 )
 
 // environmentDeployer can deploy an environment.
@@ -44,34 +41,4 @@ An environment represents a deployment stage.`,
 		"group": group.Develop,
 	}
 	return cmd
-}
-
-// bindProjectName loads the project's name to viper.
-// If there is an error, we swallow the error and leave the default value as empty string.
-func bindProjectName() {
-	name, err := loadProjectName()
-	if err != nil {
-		return
-	}
-	viper.SetDefault(projectFlag, name)
-}
-
-// loadProjectName retrieves the project's name from the workspace if it exists and returns it.
-// If there is an error, it returns an empty string and the error.
-func loadProjectName() (string, error) {
-	// Load the workspace and set the project flag.
-	ws, err := workspace.New()
-	if err != nil {
-		// If there's an error fetching the workspace, fall back to requiring
-		// the project flag be set.
-		return "", fmt.Errorf("fetching workspace: %w", err)
-	}
-
-	summary, err := ws.Summary()
-	if err != nil {
-		// If there's an error reading from the workspace, fall back to requiring
-		// the project flag be set.
-		return "", fmt.Errorf("reading from workspace: %w", err)
-	}
-	return summary.ProjectName, nil
 }
